requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
pdfplumber>=0.7.0
python-docx>=0.8.11
spacy>=3.4.0
scikit-learn>=1.0.0
requests
gunicorn